+++ changelog.md
@@ -21,6 +21,7 @@
 + Using single click payment without component now ask if you want to save it or use saved card.
 + Fixed issue where klarna status would be changed to klarna shipped and not completed when using default statuses for klarna.
 + Fixed issue where modules like VAT module would make mollie create wrong order prices. It was fixed by adding customer to context in webhook.
++ Fixed missing transaction id on order payments.
 
 ## Changes in release 4.4.3 ##
 + Bank transfer now creates order on open status.